pyproject.toml
[project]
name = "ai-agent-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.20",
    "datasets>=4.5.0",
    "fastmcp>=2.11.3",
    "ipykernel>=7.1.0",
    "litellm>=1.81.3",
    "mcp>=1.13.1",
    "openai>=1.101.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.0.0",
    "tavily-python>=0.7.11",
    "tqdm>=4.67.1",
    "wikipedia>=1.4.0",
]

[tool.setuptools.packages.find]
include = ["agent_framework*", "examples*"]
